refactored PipelineDefinitionController & test

+++ Server/src/test/java/net/hawkengine/http/tests/PipelineDefinitionControllerTest.java
@@ -1,52 +1,64 @@
 package net.hawkengine.http.tests;
 
-
 import net.hawkengine.db.redis.RedisManager;
 import net.hawkengine.http.PipelineDefinitionController;
-
+import net.hawkengine.model.ExecTask;
+import net.hawkengine.model.GitMaterial;
+import net.hawkengine.model.JobDefinition;
+import net.hawkengine.model.MaterialDefinition;
 import net.hawkengine.model.PipelineDefinition;
-
+import net.hawkengine.model.StageDefinition;
+import net.hawkengine.model.TaskDefinition;
+import net.hawkengine.model.enums.MaterialType;
+import net.hawkengine.model.enums.RunIf;
+import net.hawkengine.model.enums.TaskType;
 import net.hawkengine.services.PipelineDefinitionService;
+
 import org.glassfish.jersey.server.ResourceConfig;
 import org.glassfish.jersey.test.JerseyTest;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
-import org.omg.IOP.ENCODING_CDR_ENCAPS;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
 import java.util.ArrayList;
 import java.util.List;
+
 import javax.ws.rs.client.Entity;
 import javax.ws.rs.core.Application;
 import javax.ws.rs.core.Response;
 
 import static org.junit.Assert.assertEquals;
 
+;
+
 public class PipelineDefinitionControllerTest extends JerseyTest {
-    private static PipelineDefinitionService pipelineDefinitionService;
+    private PipelineDefinitionService pipelineDefinitionService;
     private PipelineDefinition pipelineDefinition;
+
+    @BeforeClass
+    public static void login() throws IOException, URISyntaxException {
+        RedisManager.initializeEmbededDb(6379);
+        RedisManager.connect("localhost");
+        RedisManager.startEmbededDb();
+    }
+
+    @AfterClass
+    public static void logout() throws InterruptedException {
+        RedisManager.release();
+        RedisManager.stopEmbededDb();
+    }
 
     public Application configure() {
         return new ResourceConfig(PipelineDefinitionController.class);
     }
 
-    @BeforeClass
-     public static void login() throws IOException, URISyntaxException {
-        RedisManager.initializeEmbededDb(6379);
-        RedisManager.connect("redis");
-        pipelineDefinitionService = new PipelineDefinitionService();
-    }
-
-    @AfterClass
-    public static void logout() {
-        RedisManager.release();
-    }
-
     @Test
-    public void getAllPipelineDefinition_request_emptyList(){
+    public void getAllPipelineDefinition_nonExistingObjects_emptyList()
+            throws IOException, InterruptedException {
         //Arrange
+        this.clearDB();
         List<PipelineDefinition> expectedResult = new ArrayList<>();
 
         //Act
@@ -54,29 +66,48 @@ public class PipelineDefinitionControllerTest extends JerseyTest {
         List<PipelineDefinition> actualResult = response.readEntity(List.class);
 
         //Assert
-        assertEquals(200,response.getStatus());
-        assertEquals(expectedResult,actualResult);
+        assertEquals(200, response.getStatus());
+        assertEquals(expectedResult, actualResult);
     }
 
     @Test
-    public void getById_request_pipelineDefinitionObject(){
+    public void getAllPipelineDefinition_existingObjects_allObjects()
+            throws IOException, InterruptedException {
+        //Arrange
+        List<PipelineDefinition> expectedResult = new ArrayList<>();
+        this.preparePipelineDefinition();
+        this.pipelineDefinitionService.add(this.pipelineDefinition);
+        expectedResult.add(this.pipelineDefinition);
+        this.preparePipelineDefinition();
+        this.pipelineDefinitionService.add(this.pipelineDefinition);
+        expectedResult.add(this.pipelineDefinition);
+
+        //Act
+        Response response = target("/pipeline-definitions").request().get();
+        List<PipelineDefinition> actualResult = response.readEntity(List.class);
+
+        //Assert
+        assertEquals(200, response.getStatus());
+        assertEquals(expectedResult.size(), actualResult.size());
+    }
+
+    @Test
+    public void getPipelineDefinitionById_existingObject_correctObject() {
         //Arrange
         this.preparePipelineDefinition();
         pipelineDefinitionService.add(this.pipelineDefinition);
 
-
         //Act
-        Response response = target("/pipeline-definitions/"+this.pipelineDefinition.getId()).request().get();
+        Response response = target("/pipeline-definitions/" + this.pipelineDefinition.getId()).request().get();
         PipelineDefinition actualResult = response.readEntity(PipelineDefinition.class);
 
 
-        assertEquals(200,response.getStatus());
-        assertEquals(this.pipelineDefinition.getId(),actualResult.getId());
-        this.removePipelineDefinition();
+        assertEquals(200, response.getStatus());
+        assertEquals(this.pipelineDefinition.getId(), actualResult.getId());
     }
 
     @Test
-    public void getById_getPipelieDefinitionByWrongId_errorMessage(){
+    public void getPipelineDefinitionById_nonExistingPiplineDefinition_properErrorMessage() {
         //Arrange
         String expectedResult = "PipelineDefinition not found.";
 
@@ -84,45 +115,45 @@ public class PipelineDefinitionControllerTest extends JerseyTest {
         Response response = target("/pipeline-definitions/wrongId").request().get();
 
         //Assert
-        assertEquals(404,response.getStatus());
-        assertEquals(expectedResult,response.readEntity(String.class));
-
+        assertEquals(404, response.getStatus());
+        assertEquals(expectedResult, response.readEntity(String.class));
     }
 
-
     @Test
-    public void addPipelineDefinition_onePipelineDefinition_successMessage(){
-        //Arrange
-        this.preparePipelineDefinition();
-        Entity entity = Entity.entity(this.pipelineDefinition,"application/json");
-
-        //Act
-        Response response = target("/pipeline-definitions").request().post(entity);
-        PipelineDefinition  actualResult = response.readEntity(PipelineDefinition.class);
-
-        //Assert
-        assertEquals(201,response.getStatus());
-        assertEquals(this.pipelineDefinition.getId(),actualResult.getId());
-        this.removePipelineDefinition();
-
-    }
-/*
-    @Test
-    public void addPipelineDefinition_invalidField_errorMessage(){
+    public void addPipelineDefinition_oneObject_successMessage() {
         //Arrange
         this.preparePipelineDefinition();
         Entity entity = Entity.entity(this.pipelineDefinition, "application/json");
 
         //Act
-        Response response = target().request("/pipeline-definitions).post(entity);
+        Response response = target("/pipeline-definitions").request().post(entity);
+        PipelineDefinition actualResult = response.readEntity(PipelineDefinition.class);
 
         //Assert
-        assertEquals(400,response.getStatus());
+        assertEquals(200, response.getStatus());
+        assertEquals(this.pipelineDefinition.getId(), actualResult.getId());
     }
 
-*/
     @Test
-    public void updatePipelineDefinition_existingPipelineDefinition_updatedPipelineDefinition(){
+    public void addPipelineDefinition_invalidValidation_properErrorMessage() {
+        //Arrange
+        this.preparePipelineDefinition();
+        this.pipelineDefinition.setName(null);
+        Entity entity = Entity.entity(this.pipelineDefinition, "application/json");
+        String expectedMessage = "ERROR: PIPELINE DEFINITION NAME IS NULL.";
+
+        //Act
+        Response response = target("/pipeline-definitions").request().post(entity);
+        String actualMessage = response.readEntity(String.class);
+
+        //Assert
+        assertEquals(400, response.getStatus());
+        assertEquals(expectedMessage, actualMessage);
+    }
+
+
+    @Test
+    public void updatePipelineDefinition_existingPipelineDefinition_updatedPipelineDefinition() {
         //Arrange
         this.preparePipelineDefinition();
         pipelineDefinitionService.add(this.pipelineDefinition);
@@ -134,16 +165,16 @@ public class PipelineDefinitionControllerTest extends JerseyTest {
         PipelineDefinition actualResult = response.readEntity(PipelineDefinition.class);
 
         //Assert
-        assertEquals(200,response.getStatus());
+        assertEquals(200, response.getStatus());
         assertEquals(this.pipelineDefinition.getName(), actualResult.getName());
-        this.removePipelineDefinition();
+
     }
 
     @Test
-    public void updatePipelineDefinition_nonExisting_errorMessage(){
+    public void updatePipelineDefinition_nonExistingObject_properErrorMessage() {
         //Arrange
         this.preparePipelineDefinition();
-        Entity entity = Entity.entity(this.pipelineDefinition,"application/json");
+        Entity entity = Entity.entity(this.pipelineDefinition, "application/json");
         String expectedMessage = "PipelineDefinition not found.";
 
         //Act
@@ -155,7 +186,7 @@ public class PipelineDefinitionControllerTest extends JerseyTest {
     }
 
     @Test
-    public void deletePipelineDefinition_pipelineDefintion_successMessage(){
+    public void deletePipelineDefinition_pipelineDefintionId_successMessage() {
         //Arrange
         this.preparePipelineDefinition();
         pipelineDefinitionService.add(this.pipelineDefinition);
@@ -164,11 +195,11 @@ public class PipelineDefinitionControllerTest extends JerseyTest {
         Response response = target("/pipeline-definitions/" + this.pipelineDefinition.getId()).request().delete();
 
         //Assert
-        assertEquals(204,response.getStatus());
+        assertEquals(204, response.getStatus());
     }
 
     @Test
-    public void deletePipelineDefinition_nonExistingPipelineDefinition_errorMessage(){
+    public void deletePipelineDefinition_nonExistingPipelineDefinition_properErrorMessage() {
         //Arrange
         this.preparePipelineDefinition();
         String expectedMessage = "PipelineDefinition not found.";
@@ -177,52 +208,52 @@ public class PipelineDefinitionControllerTest extends JerseyTest {
         Response response = target("/pipeline-definitions/wrongId").request().delete();
 
         //Assert
-        assertEquals(404,response.getStatus());
+        assertEquals(404, response.getStatus());
         assertEquals(expectedMessage, response.readEntity(String.class));
     }
 
 
-
-
-
     private void preparePipelineDefinition() {
+        this.pipelineDefinitionService = new PipelineDefinitionService();
         this.pipelineDefinition = new PipelineDefinition();
         this.pipelineDefinition.setName("pipelineDefinition");
-        /*
+        List<MaterialDefinition> materialDefinitions = new ArrayList<>();
+        MaterialDefinition materialDefinition = new GitMaterial();
+        materialDefinition.setType(MaterialType.GIT);
+        materialDefinitions.add(materialDefinition);
+        this.pipelineDefinition.setMaterialDefinitions(materialDefinitions);
+        StageDefinition stageDefinition = new StageDefinition();
+        stageDefinition.setName("stageDefinition");
+        JobDefinition jobDefinition = new JobDefinition();
+        jobDefinition.setName("jobDefinitio");
+        List<JobDefinition> jobDefinitions = new ArrayList<>();
+        jobDefinitions.add(jobDefinition);
         ExecTask execTask = new ExecTask();
         execTask.setName("execTask");
         execTask.setRunIfCondition(RunIf.PASSED);
         execTask.setType(TaskType.EXEC);
-        List <MaterialDefinition> materialDefinitions = new ArrayList<>();
-        MaterialDefinition materialDefinition = new GitMaterial();
-        materialDefinition.setType(MaterialType.GIT);
-        materialDefinitions.add(materialDefinition);
-        this.pipelineDefinition.setMaterials(materialDefinitions);
-        this.pipelineDefinition.setName("pipelineDefinitio");
-        List <StageDefinition> stageDefinitions  = new ArrayList<>();
-        StageDefinition stageDefinition = new StageDefinition();
-        stageDefinition.setName("stageDefinitio");
-        JobDefinition jobDefinition = new JobDefinition();
-        jobDefinition.setName("jobDefinitio");
-        List <TaskDefinition> taskDefinitions = new ArrayList<>();
+        execTask.setCommand("echo hi");
         List<String> taskArguments = new ArrayList();
         taskArguments.add("1");
         execTask.setArguments(taskArguments);
-        execTask.setCommand("echo hi");
+        List<TaskDefinition> taskDefinitions = new ArrayList<>();
         taskDefinitions.add(execTask);
         jobDefinition.setTaskDefinitions(taskDefinitions);
-        List <JobDefinition> jobDefinitions = new ArrayList<>();
-        jobDefinitions.add(jobDefinition);
         stageDefinition.setJobDefinitions(jobDefinitions);
+        List<StageDefinition> stageDefinitions = new ArrayList<>();
         stageDefinitions.add(stageDefinition);
         this.pipelineDefinition.setStageDefinitions(stageDefinitions);
-*/
 
     }
 
     private void removePipelineDefinition() {
-        this.pipelineDefinitionService.delete(this.pipelineDefinition.getId());
+        if (this.pipelineDefinitionService != null) {
+            this.pipelineDefinitionService.delete(this.pipelineDefinition.getId());
+        }
     }
 
-
+    private void clearDB() throws IOException, InterruptedException {
+        RedisManager.stopEmbededDb();
+        RedisManager.startEmbededDb();
+    }
 }
